+++ src/FarmRPGPlusError.js
@@ -1,15 +1,5 @@
 import ConsolePlus from './modules/consolePlus';
 
-class FarmRPGPlusError extends Error {
-    constructor(type, context = '', extraMessage = '') {
-        super(type || ErrorTypes.UNKNOWN);
-        this.context = context ? ` | ${context}` : '';
-        this.extraMessage = extraMessage ? ` | ${extraMessage}` : '';
-
-        ConsolePlus.error(`Error: ${this.message}${this.context}${this.extraMessage}`);
-    }
-}
-
 const ErrorTypes = Object.freeze({
     PAGE_NOT_FOUND: 'Page not found',
     PARAMETER_MISMATCH: 'Parameter type mismatch',
@@ -19,6 +9,16 @@ const ErrorTypes = Object.freeze({
     INVALID_URL: 'Invalid URL',
     UNKNOWN: 'Unknown error',
 });
+
+class FarmRPGPlusError extends Error {
+    constructor(type, context = '', extraMessage = '') {
+        super(type || ErrorTypes.UNKNOWN);
+        this.context = context ? ` | ${context}` : '';
+        this.extraMessage = extraMessage ? ` | ${extraMessage}` : '';
+
+        ConsolePlus.error(`Error: ${this.message}${this.context}${this.extraMessage}`);
+    }
+}
 
 export {
     ErrorTypes as ErrorTypesEnum,