+++ src/modules/debugPlus.js
@@ -4,7 +4,7 @@ import RouterPlus from './routerPlus';
 import StoragePlus from './storagePlus';
 class DebugPlus {
 
-    isDevelopmentMode = () => {
+    static isDevelopmentMode = () => {
         return process.env.NODE_ENV === 'development' || process.env.NODE_ENV === 'test';
     };
 
@@ -12,17 +12,21 @@ class DebugPlus {
         RouterPlus.goto(hash);
     };
 
-    applyDebugFeatures = () => {
+    static applyDebugFeatures = () => {
         $(mainView.container).on('page:init', function () {
             window.page = myApp.getCurrentView().activePage || mainView.activePage;
         });
 
+        window.clearDevLs = () => {
+            StoragePlus.clear('frpg_plus_dev');
+            console.debug('Development local storage cleared.');
+        };
+        window.StoragePlus = StoragePlus;
         window.RouterPlus = RouterPlus;
         window.goto = DebugPlus.goto;
         window.getListByTitle = getListByTitle;
-        window.StoragePlus = StoragePlus;
     };
 
 }
 
-export default new DebugPlus;
+export default DebugPlus;